# docs/02-方案/接口契约_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "安联资管运维管理平台",
    "description": "安联资管运维管理平台（恒生 O32 投资交易系统日常运维）：M1/M2/M3 核对 + 数据源管理 + 系统配置 + 任务调度中心 + 统计看板 + 数据字典 + 本地认证",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/system/subject-price-rules": {
      "get": {
        "tags": [
          "系统配置"
        ],
        "summary": "科目取价规则列表",
        "operationId": "list_rules_api_admin_system_subject_price_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectPriceRuleInfo"
                  },
                  "type": "array",
                  "title": "Response List Rules Api Admin System Subject Price Rules Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "系统配置"
        ],
        "summary": "新增科目取价规则",
        "operationId": "create_rule_api_admin_system_subject_price_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPriceRuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectPriceRuleInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/system/subject-price-rules/{rule_id}": {
      "put": {
        "tags": [
          "系统配置"
        ],
        "summary": "修改科目取价规则",
        "operationId": "update_rule_api_admin_system_subject_price_rules__rule_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPriceRuleUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectPriceRuleInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "系统配置"
        ],
        "summary": "删除科目取价规则",
        "operationId": "delete_rule_api_admin_system_subject_price_rules__rule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/params": {
      "get": {
        "tags": [
          "系统配置"
        ],
        "summary": "系统参数列表",
        "operationId": "list_params_api_admin_system_params_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SystemParamInfo"
                  },
                  "type": "array",
                  "title": "Response List Params Api Admin System Params Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "系统配置"
        ],
        "summary": "批量修改系统参数（admin）",
        "operationId": "update_params_api_admin_system_params_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemParamsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SystemParamInfo"
                  },
                  "type": "array",
                  "title": "Response Update Params Api Admin System Params Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubjectPriceRuleCreateRequest": {
        "properties": {
          "subject_prefix": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Subject Prefix",
            "description": "科目前缀（如 1101/1501）"
          },
          "price_field": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Price Field",
            "description": "取价字段（估值表列名，如 市价/单位成本）"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "科目说明（如 交易性金融资产）"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "口径提示语（报告备注用，可空）"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否启用",
            "default": true
          },
          "sort_order": {
            "type": "integer",
            "maximum": 9999.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "description": "排序号（小者先提取）",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "subject_prefix",
          "price_field"
        ],
        "title": "SubjectPriceRuleCreateRequest"
      },
      "SubjectPriceRuleInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "subject_prefix": {
            "type": "string",
            "title": "Subject Prefix"
          },
          "price_field": {
            "type": "string",
            "title": "Price Field"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_prefix",
          "price_field",
          "enabled",
          "sort_order",
          "updated_at"
        ],
        "title": "SubjectPriceRuleInfo"
      },
      "SubjectPriceRuleUpdateRequest": {
        "properties": {
          "subject_prefix": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Prefix",
            "description": "科目前缀"
          },
          "price_field": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Field",
            "description": "取价字段"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "科目说明"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "口径提示语"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled",
            "description": "是否启用"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order",
            "description": "排序号"
          }
        },
        "type": "object",
        "title": "SubjectPriceRuleUpdateRequest"
      },
      "SystemParamInfo": {
        "properties": {
          "param_key": {
            "type": "string",
            "title": "Param Key"
          },
          "param_value": {
            "type": "string",
            "title": "Param Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "param_key",
          "param_value",
          "updated_at"
        ],
        "title": "SystemParamInfo"
      },
      "SystemParamsUpdateRequest": {
        "properties": {
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Values",
            "description": "参数键值对（仅允许已登记的系统参数键）"
          }
        },
        "type": "object",
        "required": [
          "values"
        ],
        "title": "SystemParamsUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT 访问令牌",
        "scheme": "bearer"
      }
    }
  }
}